test: align test files with their source files
Each _test.go should correspond to a same-named source file (one
concept/responsibility per file + its test). Merged same-package scenario test
files (goimports reconciled imports; no duplicate-declaration risk as they
already shared the package) and renamed drifted ones. Internal (package via) vs
external (package via_test) tests kept as separate files sharing the base name.
Key moves:

  marshalevent_internal_test.go        -> stateappevents_internal_test.go
  ctxr_test.go                         -> ctx_test.go
  eventupcaster_internal_test.go       -> eventenvelope_internal_test.go
  queue_order_test.go                  -> action_test.go
  8 projector scenario tests           -> stateappevents_projector_internal_test.go
  stateappevents_runtime_test.go       -> stateappevents_projector_test.go
  sse_{brotli,liveness}_test.go        -> sse_test.go
  sse_recover_test.go                  -> recover_test.go
  backplane_{crosspod,*,reconcile,wire}_test.go -> backplane_test.go
  sess_{adopt,change}_internal_test.go -> sess_internal_test.go
  vianats conformance/embedded/epoch   -> vianats_test.go
  vianats sanitize_internal_test.go    -> vianats_internal_test.go
  picocss/maplibre/echarts/vt/showcase test renames

Left legit standalone tests (benchmarks, integration/e2e, helpers, Go examples,
cross-cutting protocol tests like shape/patch/csp_push/h-render) as-is. gofmt
clean; go vet + go test -race green across all modules.
