+++ stateappevents_projector_test.go
